requirements.txt
flask
flask-restful
flask_cors
# pdf2image
torch
torchvision
opencv-python
scikit-image
scipy
gitpython
matplotlib
numpy
Pillow==9.5.0
psutil  # system resources
PyYAML
requests
thop  # FLOPs computation
tqdm
ultralytics
faiss-cpu
six
tensorflow
vietocr
pytesseract
# Wand
PyMuPDF
openpyxl
deskew